Publish v19.0.0. [release] (#296)
Publishes a new `cimg/node` image, `cimg/node:19.0.0` specifically, to
Docker Hub for customer use. Customers using `cimg/node:current` will be
updated as well, though we can manually disable that if needed.

Rollback options:

For the main image tag, `19.0.0`, there's not really anything to roll
back. If needed, we could simply delete it. Customers would have
manually switches to this tag in order to use it in the first place.

For the alias tag `current`, it's a matter of retargeting which image it
points to. In this case, the rollback procedure would be to alias
`current` back to `18.11.0` and push that up to Docker Hub, a 1 minute
procedure.