Add arm64 support (#346)
# Description
Add arm64 support to the cimg/node image by using buildx to build multi
arch images.

The majority of the work actually lives in the cimg-shared repo which is
where the build commands are generated:

https://github.com/CircleCI-Public/cimg-shared/pull/83
https://github.com/CircleCI-Public/cimg-shared/pull/84

The main change to the Dockerfile here is an architecture check and
switch to the `USER` directive. `sudo` in buildx builds generally causes
errors when running Docker on linux.

See also changes to cimg-orb:

https://github.com/CircleCI-Public/cimg-orb/pull/51

Note: The browsers variant does not build for arm64, only amd64, due to
a limitation with selenium server. We will address this in the future,
once arm64 support is rolled out to the majority of images

Test build with buildx:
https://app.circleci.com/pipelines/github/CircleCI-Public/cimg-node/1419/workflows/daea85da-b14a-479b-adc9-d9c358135e16/jobs/1486

# Reasons
To build multi architecture images and allow cimg/node to run on arm64
natively

# Checklist

Please check through the following before opening your PR. Thank you!

- [x] I have made changes to the `Dockerfile.template` file only
- [x] I have not made any manual changes to automatically generated
files
- [x] My PR follows best practices as described in the [contributing
guidelines](CONTRIBUTING)
- [x] (Optional, but recommended) My commits are
[signed](https://docs.github.com/en/authentication/managing-commit-signature-verification/signing-commits)