Add Arm64 support (v2) (#348)
# Description
Second attempt at adding arm64 support to this image following #346
(reverted)

# Reasons
The previous method of building with buildx was causing issues with
sudo, therefore we dropped the use of sudo in favour of using `USER`
directives. This had a knock on effect of wrong permissions when it came
to using these images in a real world test.

Updates were made to the cimg orb to restart the qemu container with
privileges before the build starts, allowing us to use sudo in the
builds. This means the only changes we need to make are arch checks.

This PR also includes an updated shared module which ensures a pipeline
will fail if any image fails to build. Previously image builds could
fail silently.

# Checklist

Please check through the following before opening your PR. Thank you!

- [x] I have made changes to the `Dockerfile.template` file only
- [x] I have not made any manual changes to automatically generated
files
- [x] My PR follows best practices as described in the [contributing
guidelines](CONTRIBUTING)
- [x] (Optional, but recommended) My commits are
[signed](https://docs.github.com/en/authentication/managing-commit-signature-verification/signing-commits)